Barcode Scanner
Added functionality to scan barcodes into the membership tab. Re-used
the QR scanner. It will now scan barcodes and QR codes depending on
which screen the user comes from.

Added some potential functions to the map so that when an annotation is
tapped, the map will center and zoom on it. May get rid of the pop-up
zoomed gallery view.